+++ src/chipshot/reader/header.py
@@ -8,6 +8,7 @@ import logging
 import typing as t
 
 from .. import exceptions
+from ..config import get_config_value
 from ..shared import FileInfo
 
 log = logging.getLogger(__name__)
@@ -16,23 +17,18 @@ log = logging.getLogger(__name__)
 def handle(info: FileInfo, config: dict[str, t.Any]) -> None:
     """Detect and extract an existing header."""
 
-    suffixes = "".join(info.path.suffixes)[1:]
-    suffix = "".join(info.path.suffix)[1:]
-
-    suffixes_config = config["extension"].get(suffixes, {})
-    suffix_config = config["extension"].get(suffix, {})
-
-    style: dict[str, str] | None = config["style"].get(
-        suffixes_config.get("style", ...)
-    ) or config["style"].get(suffix_config.get("style", ...))
-
-    # If no style is defined for the file, do nothing.
-    if style is None:
+    try:
+        (style_key,) = get_config_value(config, info.path, "style")
+    except KeyError:
+        # If no style is defined for the file, do nothing.
         return
+
+    style: dict[str, str] = config["styles"][style_key]
 
     # If the first non-whitespace content doesn't start with *block_prefix*,
     # there is no header.
-    if style["block_prefix"] and not info.contents.lstrip().startswith(
+    left_stripped_content = info.contents.lstrip()
+    if style["block_prefix"] and not left_stripped_content.startswith(
         style["block_prefix"].strip()
     ):
         return
@@ -55,7 +51,7 @@ def handle(info: FileInfo, config: dict[str, t.Any]) -> None:
     # Find the end of the block prefix, if any.
     scan_start = 0
     if style["block_prefix"]:
-        scan_start = info.contents.find(style["block_prefix"])
+        scan_start = left_stripped_content.find(style["block_prefix"])
         if scan_start == -1:
             return
         scan_start += len(style["block_prefix"])
@@ -71,7 +67,7 @@ def handle(info: FileInfo, config: dict[str, t.Any]) -> None:
     # Assume that a block prefix with no prefixed lines means there is no header.
     header_end_index = 0
 
-    for current_line, next_line, index in two_lines(info.contents, scan_start):
+    for current_line, next_line, index in two_lines(left_stripped_content, scan_start):
         # Keep going if the current line starts with the line prefix.
         if current_line.startswith(line_prefix):
             header_end_index = index
@@ -96,15 +92,15 @@ def handle(info: FileInfo, config: dict[str, t.Any]) -> None:
     if not header_end_index:
         return
 
-    info.original_header = info.contents[:header_end_index]
-    if info.contents[header_end_index : header_end_index + 1] != "\n":
+    info.original_header = left_stripped_content[:header_end_index]
+    if left_stripped_content[header_end_index : header_end_index + 1] != "\n":
         msg = f"{info.path}: The header block does not end with a newline."
         log.warning(msg)
-        info.contents = info.contents[header_end_index:]
-    elif info.contents[header_end_index : header_end_index + 2] == "\n\n":
-        info.contents = info.contents[header_end_index + 2 :]
+        info.contents = left_stripped_content[header_end_index:]
+    elif left_stripped_content[header_end_index : header_end_index + 2] == "\n\n":
+        info.contents = left_stripped_content[header_end_index + 2 :]
     else:  # Only 1 newline.
-        info.contents = info.contents[header_end_index + 1 :]
+        info.contents = left_stripped_content[header_end_index + 1 :]
 
 
 def two_lines(